***** Store framebuffer data as uint32_t vectors in the render backends
This was we get 32-bit alignment for the 32-bit pixel data which matters
on x86 (unaligned 32-bit reads incur a performance penalty on x86; it
does not matter on modern ARM processors).

C++17 and beyond guarantee proper alignment for std::vector data based
on the stored elements' size.

+++ src/gui/render/opengl_renderer.h
@@ -134,10 +134,15 @@ private:
 
 	// The current framebuffer we render the emulated video output into
 	// (contains the "work-in-progress" next frame).
-	std::vector<uint8_t> curr_framebuf = {};
+	//
+	// The framebuffers contain 32-bit pixel data stored as a sequence of
+	// four packed 8-bit values in BGRX byte order (that's in memory order,
+	// so byte N is B, byte N+1 is G, byte N+2 is R).
+	//
+	std::vector<uint32_t> curr_framebuf = {};
 
 	// Contains the last fully rendered frame, waiting to be presented.
-	std::vector<uint8_t> last_framebuf = {};
+	std::vector<uint32_t> last_framebuf = {};
 
 	// True if the last framebuffer has been updated since the last present
 	bool last_framebuf_dirty = false;
